Enhancement: Rename Reporter\Reporter to Reporter\DefaultReporter and extract Reporter\Reporter interface
